+++ core/src/elements/ons-page.spec.js
@@ -217,5 +217,14 @@ describe('OnsPageElement', () => {
       expect(element.firstChild.hasAttribute('style')).to.be.true;
     });
   });
+
+  describe('autoStyling', () => {
+    it('adds \'material\' modifier on Android', () => {
+      ons.platform.select('android');
+      let e = document.createElement('ons-page');
+      expect(e.getAttribute('modifier')).to.equal('material');
+      ons.platform.select('');
+    });
+  });
 });