+++ core/src/elements/ons-tab.js
@@ -160,15 +160,14 @@ class TabElement extends BaseElement {
   createdCallback() {
     if (!this.hasAttribute('_compiled')) {
       this._compile();
-      ModifierUtil.initModifier(this, scheme);
-
-      this.setAttribute('_compiled', '');
     }
 
     this._boundOnClick = this._onClick.bind(this);
   }
 
   _compile() {
+    ons._autoStyle.prepare(this);
+
     const fragment = document.createDocumentFragment();
     let hasChildren = false;
 
@@ -197,6 +196,14 @@ class TabElement extends BaseElement {
       this._hasDefaultTemplate = true;
       this._updateDefaultTemplate();
     }
+
+    if (this.hasAttribute('ripple') && !util.findChild(button, 'ons-ripple')) {
+      button.insertBefore(document.createElement('ons-ripple'), button.firstChild);
+    }
+
+    ModifierUtil.initModifier(this, scheme);
+
+    this.setAttribute('_compiled', '');
   }
 
   _updateDefaultTemplate() {