+++ core/src/ons/notification.js
@@ -86,6 +86,8 @@ notification._createAlertDialog = function(title, message,
     result.reject = reject;
   });
 
+  modifier = modifier || dialogElement.getAttribute('modifier');
+
   if (typeof title === 'string') {
     titleElement.textContent = title;
   }
@@ -140,7 +142,7 @@ notification._createAlertDialog = function(title, message,
 
   const createButton = function(i) {
     let buttonElement = util.createElement('<button class="alert-dialog-button"></button>');
-    buttonElement.textContent = buttonLabels[i];
+    buttonElement.appendChild(document.createTextNode(buttonLabels[i]));
 
     if (i == primaryButtonIndex) {
       buttonElement.classList.add('alert-dialog-button--primal');
@@ -205,6 +207,7 @@ notification._createAlertDialog = function(title, message,
   messageElement = footerElement = null;
 
   if (modifier) {
+    dialogElement.setAttribute('modifier', '');
     dialogElement.setAttribute('modifier', modifier);
   }